refactor(data): address PR review — extract DatabaseVersioning, named params, TransactionResult.capture
- Move the DatabaseVersioning interface + its docs into its own file
  (public/database-versioning.ts), re-exported from the database barrel.
- createEmptyDatabase takes a named-object param instead of ordinal args.
- Replace the standalone captureTransaction export with TransactionResult.capture
  via the namespace pattern (co-located with the TransactionResult type); delete
  capture-transaction.ts and drop its barrel export.
