+++ packages/data/src/ecs/database/transactional-store/transactional-store.ts
@@ -12,6 +12,7 @@ import { FromSchemas } from "../../../schema/from-schemas.js";
 import { Undoable } from "../undoable.js";
 import { IndexDeclarations } from "../../store/index-types.js";
 import { PartitionKeysOf } from "../../store/partition.js";
+import { createTransactionalStore } from "./create-transactional-store.js";
 
 export interface TransactionalStore<
     C extends Components = never,
@@ -102,4 +103,32 @@ export interface TransactionResult<C = unknown> {
     // concurrency strategy / reconciler is written against.
     readonly changedComponents: Set<string>;
     readonly changedArchetypes: Set<ArchetypeId>;
+}
+
+export namespace TransactionResult {
+    /**
+     * Run `fn` against `store` as a single recorded transaction: `store` is
+     * mutated in place, and the resulting change-set is returned as a
+     * {@link TransactionResult}.
+     *
+     * The returned result is a **replicable delta**: its `redo` operations can be
+     * forwarded to a peer and replayed there with `applyOperations`, so a
+     * pluggable replication strategy can propagate an *out-of-transaction* store
+     * edit — e.g. a version-upgrade migration run inside a load handler — through
+     * whatever transport it chooses. Core stays propagation-agnostic; this only
+     * captures.
+     *
+     * The transactional wrapper is created and discarded internally, so this is
+     * the "wrap the store, do the work as one transaction, keep the change, throw
+     * the wrapper away" pattern in a single call. Ops inside `fn` must go through
+     * the provided `t` (not the outer `store`) to be recorded.
+     */
+    export const capture = <
+        C extends Components,
+        R extends ResourceComponents,
+        A extends ArchetypeComponents<StringKeyof<C>> = never,
+    >(
+        store: Store<C, R, A>,
+        fn: (t: Store<C, R, A>) => void,
+    ): TransactionResult<C> => createTransactionalStore(store).execute(fn);
 }
